requirements.txt
openai>=1.12.0
streamlit==1.39.0
pandas==2.1.4
plotly==5.18.0
pydantic==2.5.3
python-dotenv==1.0.0
openpyxl==3.1.2
Pillow==10.1.0